refactor: main content block component
